Extract global_asm module
